Extract data-provider helpers into a partial
Increment 2f of the endpoint-monolith decomposition. Moves the data-provider
workstation helper cluster (~17 methods) out of WorkstationEndpoints.cs into a new
sibling partial WorkstationEndpoints.DataProviders.cs.

Moved (contiguous closed cluster, verified 0 callers outside the block and 0 external
partial refs for every method): provider record/diagnostic builders
(BuildWorkstationDataProviderRecords, BuildFallbackDataProviderRecord,
BuildWorkstationProviderDiagnostics, BuildMetricsConnectionSummary), routing/representative
selection (ResolveRoutingBindings, SelectRepresentativeRoutingConnection,
SelectRepresentativeTrustSnapshot), provider display/capability/note resolvers, provider
trust rationale + gate impact (BuildProviderTrustRationale[FromConnection],
BuildProviderGateImpact), and provider key/score/timestamp formatting.

Behavior-preserving relocation within the single partial class: core-resident data-endpoint
callers still reach the builders across the partial, and the nested ProviderTrustRationalePayload
record stays in core (accessible cross-partial). Using directives mirror the core header to
guarantee resolution for this large move. Core drops 6798 -> 6235 lines; ratchet retightened.
